+++ R/sf_geojson.R
@@ -8,6 +8,8 @@
 #' @param simplify logical indicating if sf objects without property columns should simplify
 #' (\code{TRUE}) into a vector of GeoJSON, or return a Featurecollection with
 #' empty property fields (\code{FALSE}). If \code{atomise} is TRUE this argument is ignored.
+#' @param digits integer specifying the number of of digits to round coordinates.
+#' Defulat is \code{NULL} - no rounding
 #'
 #' @return vector of GeoJSON
 #'
@@ -45,6 +47,8 @@ sf_geojson.sf <- function( sf, atomise = FALSE, simplify = TRUE, digits = NULL )
 #' Converts `sfc` objects to GeoJSON
 #'
 #' @param sfc simple feature collection object
+#' @param digits integer specifying the number of of digits to round coordinates.
+#' Defulat is \code{NULL} - no rounding
 #'
 #' @return vector of GeoJSON
 #'
@@ -85,6 +89,8 @@ sfc_geojson.default <- function( sfc, digits = NULL ) stop("Expected an sfc obje
 #' @param simplify logical indicating if data.frame without property columns should simplify
 #' (\code{TRUE}) into a vector of GeoJSON, or (\code{FALSE}). If \code{atomise} is TRUE
 #' this argument is ignored.
+#' @param digits integer specifying the number of of digits to round coordinates.
+#' Defulat is \code{NULL} - no rounding
 #'
 #' @return vector of GeoJSON
 #'